+++ src/cypherLanguageSupport/index.ts
@@ -18,4 +18,9 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-export { Monaco, MonacoHandles, MonacoProps } from './editor/Monaco'
+export { Monaco, MonacoHandles, MonacoProps } from './monaco-editor/Monaco'
+export {
+  setupAutocomplete,
+  setEditorTheme,
+  initalizeCypherSupport as setupCypherSupport
+} from './monaco-editor/editorSupport'
