+++ src/cypherLanguageSupport/monaco-editor/CypherMonacoThemes.ts
