+++ src/cypherLanguageSupport/monaco-editor/editorSchemaConverter.ts
