+++ src/shared/modules/editor/editorDuck.ts
@@ -17,29 +17,11 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-import {
-  CypherEditorSupport,
-  EditorSupportCompletionItem
-} from 'cypher-editor-support'
-import { editor, languages } from 'monaco-editor'
+import { setupAutocomplete, setupCypherSupport } from 'cypherLanguageSupport'
 import { Action } from 'redux'
 import { Epic } from 'redux-observable'
 import Rx from 'rxjs/Rx'
 
-import {
-  monacoDarkTheme,
-  monacoLightTheme
-} from 'browser/modules/Editor/CypherMonacoThemes'
-import { CypherTokensProvider } from 'browser/modules/Editor/CypherTokensProvider'
-import cypherFunctions from 'browser/modules/Editor/cypher/functions'
-import {
-  toFunction,
-  toLabel,
-  toProcedure,
-  toPropertyKey,
-  toRelationshipType
-} from 'browser/modules/Editor/editorSchemaConverter'
-import consoleCommands from 'browser/modules/Editor/language/consoleCommands'
 import { getUrlParamValue } from 'services/utils'
 import { GlobalState } from 'shared/globalState'
 import { APP_START, URL_ARGUMENTS_CHANGE } from 'shared/modules/app/appDuck'
@@ -47,12 +29,7 @@ import {
   commandSources,
   executeCommand
 } from 'shared/modules/commands/commandsDuck'
-import {
-  DARK_THEME,
-  DISABLE_IMPLICIT_INIT_COMMANDS,
-  LIGHT_THEME,
-  OUTLINE_THEME
-} from 'shared/modules/settings/settingsDuck'
+import { DISABLE_IMPLICIT_INIT_COMMANDS } from 'shared/modules/settings/settingsDuck'
 
 export const SET_CONTENT = 'editor/SET_CONTENT'
 export const EDIT_CONTENT = 'editor/EDIT_CONTENT'
@@ -162,20 +139,6 @@ export const populateEditorFromUrlEpic: Epic<any, GlobalState> = some$ => {
     })
 }
 
-// CypherEditorSupport returns the content attributes of procedures with dots wrapped in backticks, e.g. "`apoc.coll.avg`"
-// This function strips any surrounding backticks before we use the .content value in the completion item provider
-const stripSurroundingBackticks = (str: string) =>
-  str.charAt(0) === '`' && str.charAt(str.length - 1) === '`'
-    ? str.substr(1, str.length - 2)
-    : str
-
-export const getText = (item: EditorSupportCompletionItem): string =>
-  ['function', 'procedure'].includes(item.type)
-    ? stripSurroundingBackticks(item.content)
-    : item.content
-
-const editorSupport = new CypherEditorSupport('')
-
 export const initializeCypherEditorEpic: Epic<
   { type: typeof APP_START },
   GlobalState
@@ -184,130 +147,16 @@ export const initializeCypherEditorEpic: Epic<
     .ofType(APP_START)
     .take(1)
     .do(() => {
-      languages.register({ id: 'cypher' })
-      languages.setTokensProvider('cypher', new CypherTokensProvider())
-      languages.setLanguageConfiguration('cypher', {
-        brackets: [
-          ['(', ')'],
-          ['{', '}'],
-          ['[', ']'],
-          ["'", "'"],
-          ['"', '"']
-        ],
-        comments: {
-          blockComment: ['/*', '*/'],
-          lineComment: '//'
-        }
-      })
-
-      languages.registerCompletionItemProvider('cypher', {
-        triggerCharacters: ['.', ':', '[', '(', '{', '$'],
-        provideCompletionItems: (model, position) => {
-          const { startColumn, endColumn } =
-            model.getWordUntilPosition(position)
-          const range = {
-            startLineNumber: position.lineNumber,
-            endLineNumber: position.lineNumber,
-            startColumn,
-            endColumn
-          }
-          editorSupport.update(model.getValue())
-          let items: EditorSupportCompletionItem[] = []
-          // Cypher editor support repo has internal type errors
-          try {
-            items = editorSupport.getCompletion(
-              position.lineNumber,
-              position.column - 1
-            ).items
-          } catch {}
-
-          const { CompletionItemKind } = languages
-          const completionTypes: Record<string, languages.CompletionItemKind> =
-            {
-              keyword: CompletionItemKind.Keyword,
-              label: CompletionItemKind.Field,
-              relationshipType: CompletionItemKind.Reference,
-              variable: CompletionItemKind.Variable,
-              procedure: CompletionItemKind.Function,
-              function: CompletionItemKind.Function,
-              parameter: CompletionItemKind.TypeParameter,
-              propertyKey: CompletionItemKind.Property,
-              consoleCommand: CompletionItemKind.Function,
-              consoleCommandSubcommand: CompletionItemKind.Function,
-              procedureOutput: CompletionItemKind.Operator
-            }
-
-          // word preceding trigger character, used to determine range (where to insert) procedure suggestions
-          const { word } = model.getWordUntilPosition({
-            lineNumber: position.lineNumber,
-            column: position.column - 1
-          })
-
-          const getRange = (type: string, text: string) =>
-            ['consoleCommand', 'label', 'relationshipType'].includes(type)
-              ? { ...range, startColumn: range.startColumn - 1 }
-              : ['function', 'procedure'].includes(type)
-              ? {
-                  ...range,
-                  startColumn:
-                    range.startColumn -
-                    (text.lastIndexOf(word) + word.length + 1)
-                }
-              : range
-
-          return {
-            suggestions: items.map((item, index) => {
-              const label = getText(item)
-              return {
-                label,
-                kind: completionTypes[item.type],
-                insertText: label,
-                range: getRange(item.type, label),
-                detail: item.postfix || undefined,
-                sortText: encodeNumberAsSortableString(index)
-              }
-            })
-          }
-        }
-      })
-
-      editor.defineTheme(DARK_THEME, monacoDarkTheme)
-      editor.defineTheme(LIGHT_THEME, monacoLightTheme)
-      editor.defineTheme(OUTLINE_THEME, monacoLightTheme)
-      // Browser's light theme is called 'normal', but OS's light theme is called 'light'
-      // 'light' is used when theme is set to light in OS and auto in browser
-      editor.defineTheme('light', monacoLightTheme)
+      setupCypherSupport()
     })
     .ignoreElements()
 }
-function encodeNumberAsSortableString(number: number): string {
-  // use letter prefix to encode numbers. breaks after numbers are have more than ~28 digits
-  // Monaco by default sorts its suggestion by label text. But our language support sorts
-  // after relevance first and then alphabetically.
-  const A_CHAR = 65
-  const prefix = String.fromCharCode(A_CHAR + number.toString().length)
-  return `${prefix}${number}`
-}
-
 export const updateEditorSupportSchemaEpic: Epic<Action, GlobalState> = (
   actions$,
   store
 ) =>
   actions$
     .do(() => {
-      const state = store.getState()
-      const schema = {
-        consoleCommands,
-        parameters: Object.keys(state.params),
-        labels: state.meta.labels.map(toLabel),
-        relationshipTypes: state.meta.relationshipTypes.map(toRelationshipType),
-        propertyKeys: state.meta.properties.map(toPropertyKey),
-        functions: [
-          ...cypherFunctions,
-          ...state.meta.functions.map(toFunction)
-        ],
-        procedures: state.meta.procedures.map(toProcedure)
-      }
-      editorSupport.setSchema(schema)
+      setupAutocomplete(store)
     })
     .ignoreElements()
